[TASK-202] 새로고침 수정 배포
