+++ .github/workflows/deploy.yml
@@ -3,7 +3,7 @@ name: deploy
 on:
   push:
     branches:
-      - TASK-202
+      - develop
 
 jobs:
   docker_push:
